backend/requirements.txt
# Djando
Django~=4.2
# images
pillow~=9.5.0
sorl-thumbnail~=12.9.0
# env
python-dotenv~=1.0.0
# slag
pytils~=0.4.1

# Django User Agents
django-user-agents~=0.4.0
PyYAML~=6.0
ua-parser~=0.16.1
user-agents~=2.2.0

# asinco
aiohttp~=3.8.4
aiohttp[speedups]
httpie~=3.2.1

# Deployment
uvicorn~=0.21.1
gunicorn~=20.1.0
whitenoise~=6.4.0

# Tellegram bot
aiogram~=2.25.1
uvloop~=0.17.0
ujson~=5.7.0 
aiodns~=3.0.0 
aiogram-unittest~=0.0.1

# Debagger
ipython>=8.7.0
django_debug_toolbar~=4.0.0

# lintering
flake8>=5.0.4
flake8-broken-line>=0.6.0
flake8-isort>=6.0.0
flake8-plugin-utils>=1.3.2
flake8-return>=1.2.0
pep8-naming>=0.13.3
